Server/requirements.txt
discord.py==1.2.3
paho-mqtt==1.4.0
opencv-python==4.1.0.25